requirements_sprites.txt
rembg>=0.0.50
Pillow>=10.0.0
onnxruntime>=1.16.0
